moved components to separate folder
