+++ app/javascripts/components/BountyInstruction.js
