+++ app/javascripts/components/ConferenceDetail.js
