+++ app/javascripts/components/Instruction.js
