+++ app/javascripts/components/Notification.js
