+++ app/javascripts/components/Participants.js
